Repaired errors in test...all tests pass, moved tests from compat:test to test
